FeedFetcherWorker
FeedFetcherWorker is a new async process, responsible for downloading a GTFS feed. It can run daily and populate the FeedVersion records in the Datastore.

The "meat" of the GTFS import process now lives in FeedEaterWorker (rather than FeedEaterFeedWorker). A FeedEaterWorker is triggered for one FeedVersion model. That means that FeedFetcherWorkers should run first to populate all the local FeedVersion's. Afterwards, a FeedEaterWorker can be run on a given FeedVersion (specified by its SHA1) or it can be run on the most recent FeedVersion.

TODO: Figure out the sort order on FeedVersion's.