reduce Sidekiq concurrency on dev stack
We've been running out of database connections on the dev stack, probably because we're spinning up multiple background worker nodes. If each has 20 DB connections, that will max out the small RDS instance on the dev stack.

TODO: In the future, figure out the appropriate concurrency number for production. Maybe it has to be dynamically dependent on the number of worker nodes.