fix Sidekiq middleware changes
The web dashboard has been internally refactored in SIdekiq 4.2.0. It now uses straight Rack, rather than Rack and Sinatra: https://github.com/mperham/sidekiq/blob/master/Changes.md#420

Removing a hack that was useful in the past but no longer works (and now causes the error that led to having to revert in #810).